Add a DragonFly implementation
This follows a similar pattern to FreeBSD, but without the embedded size
check in the struct.  It currently depends on an unreleased version of
libc so is not yet suitable for merging.

Since the struct is not zero-initializable due to some embedded enums,
this requires the use of MaybeUninit, raising the MSRV to 1.36.  This
seems reasonable given 1.28 is nearly four years old.

Due to DragonFly's lazy reaping of LWP's after thread exit, the tests
now wait up to 50ms for the thread count to reach 1 - also providing a
useful place to warn about null implementations and --test-threads=1.