Add an OpenBSD implementation
This is somewhat more complex than FreeBSD and DragonFly because there
is no dedicated thread count field in kinfo_proc - we must request a
list of threads, making two calls to sysctl so we know how big a buffer
we should allocate, and then counting the results of a second call.

The struct is currently hardcoded and needs upstreaming to libc.