Rewrite tests
Previously a macro was used to recursively spawn threads to perform
asserts within - meaning failures involved confusing and ugly chains of
thread join panics.

Replace this with a function that spins up a given number of threads and
calls a closure on the main thread.  This leads to much tidier failure
output.

Add some feedback for which test is running to aid in debugging
failures.